requirements.txt
asgiref==3.2.10
certifi==2020.6.20
chardet==3.0.4
Django==3.1.2
django-simple-sso==0.14.1
idna==2.10
itsdangerous==1.1.0
pytz==2020.1
requests==2.24.0
sqlparse==0.3.1
urllib3==1.25.10
webservices==0.7
